backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic>=2.7.0,<3.0.0
openai==1.66.3
httpx==0.28.1
pytest==8.3.4
